backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic[email]==2.9.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
python-dotenv==1.0.1
cloudinary==1.41.0
apscheduler==3.10.4
qrcode==7.4.2
pillow==10.4.0
